面值更改

+++ czb-global/src/main/resources/db/migration/V9.2__add_czb_equity_payment.sql
@@ -9,7 +9,7 @@ CREATE TABLE `czb_equity_payment_commodity` (
     `purchase_price` double(10,2) NULL DEFAULT NULL COMMENT '进价',
     `selling_price` double (10,2) NULL DEFAULT NULL COMMENT '售价',
     `goods_pic` varchar(50) NULL DEFAULT NULL COMMENT '商品图片',
-    `unit_price` int(8) NULL DEFAULT NULL COMMENT '商品面值',
+    `unit_price` double(10,2) NULL DEFAULT NULL COMMENT '商品面值',
     `is_game` int(1) NULL DEFAULT NULL COMMENT '是否是游戏（为1时需要传入区服信息）',
     `desc` varchar(1000) NULL DEFAULT NULL COMMENT '描述信息',
     `on_sale` int(1) NULL DEFAULT NULL COMMENT '是否开售',
@@ -68,7 +68,7 @@ CREATE TABLE `czb_equity_payment_order` (
     `code` varchar(100) NULL DEFAULT NULL COMMENT '卡密 / 链接',
     `purchase_price` double(10,2) NULL DEFAULT NULL COMMENT '进价',
     `selling_price` double (10,2) NULL DEFAULT NULL COMMENT '售价',
-    `unit_price` int(8) NULL DEFAULT NULL COMMENT '商品面值',
+    `unit_price` double(10,2) NULL DEFAULT NULL COMMENT '商品面值',
     `create_at` timestamp NULL DEFAULT NULL COMMENT '创建时间',
     `update_at` timestamp NULL DEFAULT NULL ON UPDATE CURRENT_TIMESTAMP COMMENT '更新时间',
     PRIMARY KEY (`order_id`) USING BTREE
